removed souts and some refactoring
